add para o heroku

+++ src/main/java/br/com/firzen/campeoanto/model/Role.java
@@ -1,8 +1,11 @@
 package br.com.firzen.campeoanto.model;
 
-import java.util.*;
-
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -15,10 +18,10 @@ public class Role {
     @Id
     @Column(name = "role_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
      
     private String name;
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
